feature(Dashboard): Add news widget to all dashboards
Because it meant a lot to the marketing team.

The marketing team also wanted the news feed to always be visible (i.e.
users should not be able to remove it from their views).

Had to revamp the html stripper because the old one tried to download
all images from the news feed. But that resulted in 404 since the image
URLs are relative. It was a bad idea anyway to download the images since
we aren't going to show them.

ref: UC-846
code-review: #4178