fix(TOF): tidy toolbar + add link and print badge to customer card
Because toolbar dont need customer name when we have a customer card,
so link was moved to customer card. We dont need sum when we have it in
a footer. We dont need share status when we have share badges in
customer card, so missing print badge was added to customer card.

Ref: UK-972
PR: #4190