Bug/dh/uk920 payment modal (#4191)
bug(Payment): fixed the addPaymentmodal

The paymentmodal was not working anymore. The problem was that it added a bankaccount while saving. This was also done by the bankaccount modal. This resulted in an error, since both tried to save the same bankaccount. The bankaccount modal will from now on only save the bankaccount if a boolean value _saveBankaccountInModal is passed and set to true. Because is some cases you want to change the bankaccount directly from the modal.
Besides that I also changed how the modal looks and works:
Changed from bankaccount to a dropdownControl. This forces the user to select a valid from bankaccount
Changed the toBankaccount into a multiview so the user can add a missing bankaccount.
Added validation to the modal so the user is forced to enter a frombankaccount, tobankaccount, amount , paymentdate and a valid relation.
Besides that the autocomplete is changed. it now returns to the last selected value if nothing is found

ref: UK-920