hotfix: make sure bank users always have at least one role
This fixes an issue where setting users without roles as bank user
causes them to lose most of the permissions required to make the
system work, in worst case scenarios locking themselves out of the
system. This should be fixed on backend, but for now a frontend
hack will do.

UC-892