fix(Bill): Setting inbox preview to read-only
Because now you can upload more images to an inbox item, and since an
inbox item _is_ a file, it doesn't make sense to allow that.

A small caveat is that with the new files API we _can_ actually upload
multiple files to one file with some join magic, but since implementing
that is beyond the scope of this task I'm just going to make it read
only.

There is another bug that removes thumbnails if you make a uni-image
component read-only. I have created the issue UC-939 for that.