feat(Marketplace): Remove "allerede" wording form "has bought" text
Because it made it look like you had already bought a product when you
pressed the buy button, now it just says a more neutral "You have bought
<product name>".

Also changed the link from settings/company --> settings/users, because
that feels more natural even though that is not the final screen that
the user is supposed to be taken to in the end. The is going to be a new
view that the users should be sent to. But this is better in the mean
time.

ref: UC-859