fix(journalentry): avoid timeout bug
The setJournalEntryData function is setting a timeout of 500ms if
vattypes does not exist. The problem is that this function is called
multiple times, first with empty array, then with the actual lines
(if there are any).

Since vattypes does not exist when calling setJournalEntryData
with empty array the function does a timeout of 500ms then tries
setting the data again. In the meantime the actual lines are set,
and since the vattypes request has finished this is done instantly.

When the initial timeout then completes it overrides the lines to
empty array.

By only calling loadData() in ngOnChanges if vattypes exist we avoid
triggering the timeout all together.

Hopefully..