fix(accounting): fix problem in bill view when splitting a file
because the changes are now not detected so when saving a new supplier invoice, the incorrect file will be referenced (the old file id, before splitting). Also make some adjustments to the hasChangedFiles function, it compares by "id", but the field is named "ID" so changes are never detected it seems.

ref UK-1208