fix(accounting): dont show credited journalentries when entering invoice number
credited posts is currently displayed because the restamount is not set to 0. But it should never be showed, so just filter for the "open" and "partially marked" statuses when getting journalentrylines based on the entered invoice number. Should consider blocking this in the API as well, but this solves the immidiate problem

ref SAK-1152