fix(uni-image): avoid retry loop in img element error handler
The error handler was not checking if re-auth had already been
attempted, so an incorrect reference to the file would lead to an
infinite loop of re-auth requests.

The retry logic on file requests is still sort of fragile, because
it relies on us not messing up the value of the class variable
didTryReAuthenticate. However its the best we can with the current
solution of having a separate files token. Reason being that the 401
does not actually happen in a GET request we have control over,
but in the request "initiated" by <img src="url with files token"/>