fix(accountSettings): select update value when model changes
Select didn't update the selected value if we change the model and reuse the component. It was because when we emit the change value from the select the wrapper didn't update the input value. That means that when the new model comes the value detected by the component was the last value, and the wrapper didn't notify the component to update the input. Now when the select changes the value the wrapper also does it. So the value is updated.