fix(uniform): move focus to the right component when user clicks in a field after a model change
The error was that when user updates a field and then click on other field focus returned to the edited field. That was because out of the form we emit a change to the form and focus handler had two fields asking for the focus: one was the field user click and other was the previous field saying uniform to refresh and that means to lose the focus, then the edited field ask again for the focus. By timing always the edited field won.