fix(EmployeeLeave) fix delete problem (#4480)
this happened because we didn't set the element in employee leave list to the new saved leave. Corrected this by using list[index] = item instead of elementInList = item

migrated to ag-grid.
also added double binding to the table. This way our list is always in sync with ag-grid, which again simplified our delete functionality.

UL-669