fix(Sales): Rename sales/InvoiceTypes.ts --> sales/invoiceTypes.ts
Because that is the naming standard we have and that is how the imports
of that file is written. Writing the imports in the same letter case as
the is now vitally important after we started using angular-cli since it
will break on a mac/linux pc if it detects a difference in letter
casing.