feature(WageType) disable HideFromPaycheck if tax or payment (#4509)
because we don't want to hide transes on given wagetype if it is subject to tax draws or is marked as payment.
also made getDefault into a pure function. And removed get default in child view. The functionality had been moved to header components default method.
So we have everything in one place.

UL-696