fix(customer/supplier): multiple fixes to canDeactivate
- The previous canDeactivate function tried to check CountryCode
without proper null guarding. This resulted in the application
crashing if you tried exiting a supplier/customer without an
address..

- Rewrite the check to look at invoiceAddress. Previously it just
checked the first address in the addresses array. This might not
even be in use, and is a pretty bad way to check if the customer or
supplier is foreign.

UK-1356