fix(jobs): use queryParams instead of params on details route
Because the param jobName is a string, and since our authGuard
has no way of knowing that the last par of for example
"/job-logs/myjobname" is a param it cannot build a correct
permission string for it.

This is just a quick fix for the issue. Will look further into this
one as I'm not quite happy about the routing

UK-1377