feat(numberInput): rework number formatting/parsing in uniform
Previously the form input would round anything that went into the
field down to two (or whatever is in company settings) decimals,
both when formatting display value and when parsing an input value.

The problem with this is that backend does not have this decimal
limitation, leading to cases where the backend value is not the
same as what we display on frontend. A real life example of where
this is an issue is a product with a unit price of say 0.298.
Frontend would show 0.3 here, which is not correct.

We might need to do a more elegant solution for handling numbers
in general later, but for now this commit should solve the reported
problems. TLDR of the changes is that we only add "missing"
decimals, never round or remove extra decimals from the number.

E.g with company settings specifying two decimals:
2.5 becomes 2.50, but 2.475 does not become 2.50

(You'll still see rounding happening in the application, but it
should be an intended action, not something that happens by default)

To avoid breaking the entire world this is only applied to
format 'money', at least for now, as this is where we have
gotten complaints

UC-1093