fix(filters): remove searchValue field and move date autocompletion
Because the latest fix to date autocompletion on filters caused
pre-defined config filters to stop working. I've moved some logic
around to fix this. The searchValue field on ITableFilter has been
removed, "value" is now used for both the filter input and the
actual value that goes into the filter string.

The logic for autocompleting date on filter operators that require
a valid date string has been moved to getFilterValueFromFilter(..)
so that it only affects the value used when filtering, not the
filter input.

UPB-295