feat(distribution): refactor distribution to be handled by job server
remove some of the options to distribute an invoice, this should be handled by just sending "distribute" and let the jobserver handle the rest. Should consider removing Send EHF as well, but this is currently used to activate EHF - so it might be better to handle this in the distribution tab (e.g. not allow setting a distributionplan that prefers EHF unless EHF is activated, and place the button there). The same might be done in the distribution plan setup for customer and companysettings