refactor(productgroups): move away from angular-tree-component
Because its no longer mantained and wont work with the angular
version we want to update to.
Also made some minor improvements to the ui/flow, but note that this
is mostly a refactor, and the productgroups view is still not great.

The økonomi team should consider going over and improving this view
when they have time.

UC-1254