fix(accounting): transquery filter
use text input instead numeric

this is because numeric has a concrete behaviour in uniform. In uniform
we expect to write a value and then jump to the next element. In that
moment uniform checks numeric value and cleans it. In that case we want
numeric to be an instant input. that implies that if we are writing a value
and it is not valid uniform is going to clean it.

i.e: I write '-' so it is not a number so uniform cleans it.
other i.e: I write '1,5'. In javascript comma separator is dot, so it is
not taken as valid number.