hotfix(marketplace): fix incorrect activation check on extensions
Incorrect use of "magic strings" when checking if an extension was
properly activated.

I've rewritten the activation check in EHFService so this doesn't
happen again. It should work exactly like before, but now has
constant naming instead of passing strings to the function.

Since I had to rewrite parts of it anyways I've also moved the
activation check out of subscribe modal into the modules view.
This allows us to not show the green check mark until the extension
is completely activated.