Bug (JournalEntryService) Agio Account bug and sum difference (#4848)
When using the manual journalentry module and selecting a customer invoice with foreign currency as a counter post the agioline was wrong.
It had a duplicate account. fix; Removed duplicate account.

Calculation of exchangerate was wrong. Fix: Routine is updated and deducts bankcharges from amount

Bankcharges line was missing even when these were specified.
Fix: added a bankcharges line

onChange Event handler for modal was not working. fixed this.

Ref: CM-97
PR: #4848