fix(ag-grid): fix edge case error in grid layout calculation
Due to a timing issue when loading data into infinite scroll tables
the grid would sometimes make room for scrollbars that weren't
actually needed. This caused "gaps" in the layout, and in some cases
made the grid calculate col width incorrectly.
Pushing the doLayout call to the end of the call stack with an empty
timeout allows the rows to be rendered properly before the layout
is calculated.