fix(supplierInvoice): on draft lines set deductions to 0 when value is null
on supplierinvoice when user set a new draft line that has a default deduction value it can't be set to null. In case you set it to null, draft line will get the default value again.

To solve that we set value to 0 in case we have null as value of VatPercentDeductions property.