fix(uniform): fix navigation when 2 fields have the same property
it happens 2 problems. First problem is about event. uniform didn't check if event is null. Now it does. Second problem is uniform can't find a field if there is other field with the same property but earlier in the array of fields. To solve that I have added a new property to find it: the label.