feat: rewrite/add user management, elsa, contract-activation
- Rewrite user management. Use new elsa endpoint and implement
new design on views and dialogs

- Rewrite elsa purchases in frontend. Consolidate the old service
functions and types (there were many..) into a single way to get
and put purchases.

- Move elsa models to @app/models instead of @app/services.

- Use new purchases endpoint in marketplace. Misc tweaks to styling
to make it more in line with design sketches.

- Move autobank agreement dialog to a common module and use as
activation funtion in marketplace.

- Add view for contract activation. Due to the backend endpoint not
being done in time this view wont be used yet, but is more or less
ready to go once we hook it up with the backend.

SW-1 SW-29, SW-30 SW-85