refactor(year): remove accountingyear companysettings and improve local year handling
use localStorage to keep track of active year per company instead of relying on companysettings serverside because the existing solution caused a lot of problems last year. UE will now prompt user to switch year when entering a new year, and save this info in localStorage. This way users can stay in the year they want. Frontend will now also ask the API to actually create the selected year, instead of just setting it in localStorage and create it e.g. when the first invoice is invoiced.

Also make the datepicker a bit more "context aware"; if the user is in january in the current year (i.e. calendar year), and enters e.g. 1512, he/she most likely means 15.12 _last year_. Use a floating month interval, so e.g. 1510 will translate to 15.10.<lastyear> in january, but 15.10.<thisyear> in april. 

ref #CM-101