fix(accounting): avoid problem with nondistinct rows in supplierinvoice ticker
the duplicates were caused by the join on task/approvals when the approvers were not included in the select. This caused duplicates records, and because of skip the distinct does not work quite as intended in statistics. Fixed by always getting the stuffed approverlist, even if it is not displayed

ref #CM-250