feat(files): change tokenhandling for postrequests
because posting data might as well use the real UE token instead of the Uni Files token - this simplies authentication and the code, and also solves problem with some of the requests not implementing retry logic (e.g. the sync method). This will also reduce the number of 401 errors in unifiles

also refactor the rest of the uploadcomponents to use UE token as well

ref #CM-349