feature(ControlModal) update
Clean up sums. Use sums component instead of uniform, because uni-form is overkill for readonly values. 
Also moved over to a single grid with grouping. This needed some changes in ag-grid in collaboration with @urrang, because we needed to be able to controll the summing(only sum transes that has bas_payment set to true).

MIN-288