update(Transquery): Update on linking to view + keep state
Let "Show credited lines" checkbox be always visible in the transquery details. Default hidden credited lines in the journalentrylines ticker, and new checkbox that to show hidden lines.

Major update on linking to transquery details and keeping the search fields.. Fill in the fields from the searchparams and keep state when changing. Tweek ticker linking to go to
transquery view when clicking the JournalEntryNumber, not journalentrymanual view as before..

ref: CM-383