fix(ag-grid): call dataMapper if defined and remove old mappers
We originally made the dataMapper for mapping statistics data
objects to arrays. Since this is done automatically in ag-grid
we never bothered calling the function there.
Since we function can be useful for other things I've added it
to ag now, and removed the old mappers that were made for unitable