fix(supplierinvoice): set account properly to allow open accrual modal
accrual modal was not able to be opened if account on draft lines come from
smart booking. That was because the source of smart booking gets account with
less information than the autocomplete search function. Now the information that
was missed in the smart booking function is added in the http call.

closes: #5490 
ref: TC-738