featue(Employee) only show employees with transes on run
this makes the payroll run much easier to use on companies with over 100 employees. Showing only what employees that are there. Also the only category filter here is when we search for new employees. We then filter out employees already added and not in category selection.

We also use virtual scroll, so we don't get performance issues with many employees. The reason we didn't go for infinite scroll (pulling employees from backend when we scroll down to them), is because we also wan't to save some state about what employee had focus, what sorting we have and what search we have.
MIN-317