fix(employees) various fixes on payrollrun emps
* fix(employees) fix sorting

now we reset the other sortings when we switch. This way the user will get the sorting they expect when switching back and forth.
Also switched around so we filter first and sort after. This way we get a copy of the list from filter function and also if we search, we only sort on the filtered list, meaning less work.

* fix(EmployeeFilter) fix filter when there are only one fromTo

* fix(PayrollRun) add refresh emp action

because we need to load employees again if we get transes from other places (hours, vacation).

* fix(EmployeeModal) exclude already chosen employees

also removed this.options.data.employee because it is exactly the same as this.employees

MIN-317