fix(Employee) fix minor bugs on payrun
* fix(Employee) remove emp when pressing X

the method was not implemented

* fix(EmployeeSorting) fix error sorting

it did not use errors map, and it needed to be converted to 1 or 0 because it either is an error or not. Otherwise the sorting wouldn't work

* fix(Employee) fix exclude

this error made us exclude one to many ids. We had employee 1,2 and 3 but wanted to search for 4. Since 4 is the highes number you get from the exclude fromtoos we excluded that one too for good measure

MIN-317