fix(Regulative) remove product check (#6162)
because if the call is forbidden, we already know that the user doesn't have the product. We don't know how many bundles this product will be in. If it is REGULATIV, LONN_UTVIDET or SOME_FUTURE_BUNDLE. This makes our life easier by just checking if the call is forbidden or not.

MIN-575