update(BankReconciliation): Updates on bank rec journaling modal (#6562)
When opening model with lines, autorun rules in the back and show hint of result if rules match. Flash lines that were updated in the grid. Fix small bug where numberseries in session is stored from previous company when switching company or year. Also added to default rules to the view that users can click and collect.. You now get full lines text on hover on statement lines

PR: #6562