fix(BankRec): Fix autoaddrow bug + no vat lines fix (#6741)
* fix(Autobank): SR always has autobank

When in SR environment, always allow admins to create bank users

ref: SR-203

* fix(BankRec): Fix autoaddrow bug + no vat lines fix

The files field caused the keyboard navigation to fail, and a new row was never added to. Added kayboardnavigationskip to the field in the tableconfig. Also fixed case where the user removed the vatcode from the line, it was reentered before saving based on either creditaccountvattype or debitaccountvattype. If the user now removes the vat, it has to be manually but back

ref: SR-166 BN-780
PR: #6741